# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PartSignal API
  version: 0.1.1
  description: PartSignal MVP 契约。除登录外，写请求必须携带 X-CSRF-Token。
servers:
  - url: /
security:
  - sessionCookie: []
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200':
          description: 登录成功并设置会话 Cookie
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthSession'}
        '401': {$ref: '#/components/responses/ErrorResponse'}
  /api/v1/auth/logout:
    post:
      tags: [auth]
      operationId: logout
      parameters: [{$ref: '#/components/parameters/CsrfHeader'}]
      responses:
        '204': {description: 已撤销会话}
        '401': {$ref: '#/components/responses/ErrorResponse'}
  /api/v1/auth/me:
    get:
      tags: [auth]
      operationId: getCurrentUser
      responses:
        '200':
          description: 当前用户
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
        '401': {$ref: '#/components/responses/ErrorResponse'}
  /api/v1/auth/csrf:
    get:
      tags: [auth]
      operationId: getCsrfToken
      responses:
        '200':
          description: 当前会话 CSRF Token
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CsrfToken'}
        '403': {$ref: '#/components/responses/ErrorResponse'}
  /api/v1/auth/change-password:
    post:
      tags: [auth]
      operationId: changePassword
      parameters: [{$ref: '#/components/parameters/CsrfHeader'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChangePasswordRequest'}
      responses:
        '204': {description: 密码已修改且其他会话已撤销}
        '401': {$ref: '#/components/responses/ErrorResponse'}
  /api/v1/users:
    get:
      tags: [identity]
      operationId: listUsers
      responses:
        '200':
          description: 用户列表
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserList'}
    post:
      tags: [identity]
      operationId: createUser
      parameters: [{$ref: '#/components/parameters/CsrfHeader'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UserCreate'}
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
        '409': {$ref: '#/components/responses/ErrorResponse'}
  /api/v1/users/{user_id}:
    patch:
      tags: [identity]
      operationId: updateUser
      parameters:
        - name: user_id
          in: path
          required: true
          schema: {type: string, format: uuid}
        - {$ref: '#/components/parameters/CsrfHeader'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UserUpdate'}
      responses:
        '200':
          description: 已更新用户账号类型或启用状态
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
        '409': {$ref: '#/components/responses/ErrorResponse'}
  /api/v1/users/{user_id}/reset-password:
    post:
      tags: [identity]
      operationId: resetUserPassword
      parameters:
        - name: user_id
          in: path
          required: true
          schema: {type: string, format: uuid}
        - {$ref: '#/components/parameters/CsrfHeader'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ResetPasswordRequest'}
      responses:
        '204': {description: 已设置临时密码并撤销目标用户会话}
        '409': {$ref: '#/components/responses/ErrorResponse'}
  /api/v1/audit-logs:
    get:
      tags: [identity]
      operationId: listAuditLogs
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - {name: target_type, in: query, schema: {type: string}}
        - {name: target_id, in: query, schema: {type: string, format: uuid}}
      responses:
        '200':
          description: 追加式审计日志
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuditLogList'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: partsignal_session
  parameters:
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: true
      schema: {type: string, minLength: 32}
...
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    PageSize:
      name: page_size
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
...
  responses:
    ErrorResponse:
      description: 业务或校验错误
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: {$ref: '#/components/schemas/ErrorDetail'}
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code: {type: string}
        message: {type: string}
        details: {type: object, additionalProperties: true, default: {}}
        request_id: {type: string}
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: {type: string, minLength: 1}
        password: {type: string, minLength: 8}
    AuthSession:
      type: object
      additionalProperties: false
      required: [user, csrf_token]
      properties:
        user: {$ref: '#/components/schemas/User'}
        csrf_token: {type: string}
    CsrfToken:
      type: object
      additionalProperties: false
      required: [csrf_token]
      properties:
        csrf_token: {type: string}
    AccountType:
      type: string
      enum: [ADMIN, ENGINEER]
    User:
      type: object
      additionalProperties: false
      required: [id, username, display_name, account_type, is_active, must_change_password, revision, created_at]
      properties:
        id: {type: string, format: uuid}
        username: {type: string}
        display_name: {type: string}
        account_type: {$ref: '#/components/schemas/AccountType'}
        is_active: {type: boolean}
        must_change_password: {type: boolean}
        revision: {type: integer}
        created_at: {type: string, format: date-time}
    UserCreate:
      type: object
      additionalProperties: false
      required: [username, display_name, password, account_type]
      properties:
        username: {type: string, minLength: 3}
        display_name: {type: string, minLength: 1}
        password: {type: string, minLength: 12}
        account_type: {$ref: '#/components/schemas/AccountType'}
    UserUpdate:
      type: object
      additionalProperties: false
      required: [expected_revision, display_name, account_type, is_active]
      properties:
        expected_revision: {type: integer, minimum: 0}
        display_name: {type: string, minLength: 1}
        account_type: {$ref: '#/components/schemas/AccountType'}
        is_active: {type: boolean}
    ResetPasswordRequest:
      type: object
      additionalProperties: false
      required: [temporary_password]
      properties:
        temporary_password: {type: string, minLength: 12}
    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [old_password, new_password]
      properties:
        old_password: {type: string, minLength: 8}
        new_password: {type: string, minLength: 12}
    UserList:
      type: object
      additionalProperties: false
      required: [items, page, page_size, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/User'}}
        page: {type: integer}
        page_size: {type: integer}
        total: {type: integer}
    AuditLog:
      type: object
      additionalProperties: false
      required: [id, actor_id, action, target_type, target_id, change_summary, request_id, created_at]
      properties:
        id: {type: string, format: uuid}
        actor_id: {type: string, format: uuid}
        action: {type: string}
        target_type: {type: string}
        target_id: {type: string, format: uuid}
        change_summary: {type: object, additionalProperties: true}
        request_id: {type: string}
        created_at: {type: string, format: date-time}
    AuditLogList:
      type: object
      additionalProperties: false
      required: [items, page, page_size, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/AuditLog'}}
        page: {type: integer}
        page_size: {type: integer}
        total: {type: integer}
...